pyproject.toml
[tool.poetry]
name = "mmr"
version = "0.1.0"
description = ""
authors = ["Pedro Rodriguez <me@pedro.ai>"]

[tool.poetry.dependencies]
python = "3.9.*"
typer = "^0.4.0"
luigi = "^3.0.3"
pandas = "^1.3.4"
pydantic = "^1.8.2"
numpy = "^1.21.4"
streamlit = "^1.2.0"
watchdog = "^2.1.6"
pysimdjson = "^4.0.2"
altair = "^4.1.0"
altair-saver = "^0.5.0"
pedroai = "^0.3.0"
torch = "^1.11.0"
krippendorff = "^0.5.1"
jupyterlab = "^3.4.0"
rbo = "^0.1.2"
tabulate = "^0.8.9"
rich = {extras = ["jupyter"], version = "^12.4.1"}
python-Levenshtein = "^0.12.2"
scikit-learn = "^1.1.0"
transformers = "^4.19.0"
pyfunctional = "^1.4.3"
pandasql = "^0.7.3"

[tool.poetry.dev-dependencies]
ipython = "^7.30.0"
black = "^21.11b1"
pylint = "^2.12.1"
jupyter = "^1.0.0"
jupyter-http-over-ws = "^0.0.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
